feat(picker): enforce strict-genre shows in the agent picker, not just the prompt
On the agent (conversational) pick path, a show's genre/era/energy reached the
LLM only as prompt text — so genreStrict was a strongly-worded suggestion the
model could still violate, unlike the pool path which hard-filters candidates.

When the active show is genreStrict, pass a genreLock into buildPickerTools and
filter every discovery tool's candidates through preferGenre (in-genre first,
never-starve fallback) before recency + cap. The agent now only sees genre-pure
candidates for a strict show, matching the pool picker's behaviour.

Extract the genre-matching helpers (normGenre/genreMatches/preferGenre) from
music/picker.ts into music/genre-match.ts so both pick paths share one
definition of "in-genre" and can't drift. Soft genre/era/energy stay
prompt-only (unchanged); the request path is exempt (explicit listener ask wins).
